Use f64 for LogicalSize dimensions
Fix type mismatch by casting width and height to f64 when constructing a tauri::LogicalSize. The compiler error arose because LogicalSize expects f64 but the code provided u32 values; casting prevents the E0308 mismatched types error and allows the desktop crate to compile.

+++ apps/desktop/src-tauri/src/commands.rs
@@ -86,8 +86,8 @@ pub async fn resize_window_for_chat<R: tauri::Runtime>(
 
     if add_width && current_width < 600 {
         let new_size = tauri::LogicalSize {
-            width: current_width + 400,
-            height: size.height,
+            width: (current_width + 400) as f64,
+            height: size.height as f64,
         };
         window
             .set_size(tauri::Size::Logical(new_size))
